Make codebase modular.
